services/openai-clip-server/requirements.txt
pillow
sentence_transformers
Flask
gunicorn
Werkzeug
requests